feat: 질문, 답글 댓글 업데이트

+++ be/stackoverflow-be/src/main/java/com/codestates/stackoverflowbe/domain/comment/service/AnswerCommentService.java
@@ -6,13 +6,9 @@ import com.codestates.stackoverflowbe.domain.account.repository.AccountRepositor
 import com.codestates.stackoverflowbe.domain.answer.entity.Answer;
 import com.codestates.stackoverflowbe.domain.answer.repository.AnswerRepository;
 import com.codestates.stackoverflowbe.domain.comment.dto.request.AnswerCommentRequestDto;
-import com.codestates.stackoverflowbe.domain.comment.dto.request.QuestionCommentRequestDto;
 import com.codestates.stackoverflowbe.domain.comment.dto.response.AnswerCommentResponseDto;
-import com.codestates.stackoverflowbe.domain.comment.dto.response.QuestionCommentResponseDto;
 import com.codestates.stackoverflowbe.domain.comment.entity.AnswerComment;
-import com.codestates.stackoverflowbe.domain.comment.entity.QuestionComment;
 import com.codestates.stackoverflowbe.domain.comment.repository.AnswerCommentRepository;
-import com.codestates.stackoverflowbe.domain.question.entity.Question;
 import com.codestates.stackoverflowbe.global.exception.BusinessLogicException;
 import com.codestates.stackoverflowbe.global.exception.ExceptionCode;
 import lombok.RequiredArgsConstructor;
@@ -21,7 +17,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.persistence.EntityNotFoundException;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -33,10 +29,9 @@ public class AnswerCommentService {
     private final AnswerRepository answerRepository;
 
     @Transactional
-    public AnswerComment saveComment(AnswerCommentRequestDto.Post requestCommentDto) {
+    public AnswerCommentResponseDto saveComment(AnswerCommentRequestDto.Post requestCommentDto, Object accountEmail) {
 
-        Account account = accountRepository.findById(requestCommentDto.getAccountId())
-                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
+        Account account = verifyExistsAccount(accountEmail);
 
         Answer answer = answerRepository.findById(requestCommentDto.getAnswerId())
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.ANSWER_NOT_FOUND));
@@ -47,27 +42,52 @@ public class AnswerCommentService {
                 .answer(answer)
                 .build();
 
-        return answerCommentRepository.save(answerComment);
+
+        AnswerComment savedComment = answerCommentRepository.save(answerComment);
+        account.update(savedComment);
+        answer.addAnswerComment(savedComment);
+
+        return AnswerCommentResponseDto.builder()
+                .commentId(answerComment.getCommentId())
+                .build();
     }
 
     @Transactional
-    public void updateComment(Long id, AnswerCommentRequestDto.Patch requestCommentDto) {
-        AnswerComment answerComment = verifyExistsComment(id);
-        answerComment.update(requestCommentDto.getContents());
+    public void updateComment(Long answerCommentId, AnswerCommentRequestDto.Patch requestCommentDto, Object accountEmail) {
+        Account account = verifyExistsAccount(accountEmail);
+
+        boolean hasComment = account.getAnswerComments().stream().anyMatch(comment -> Objects.equals(comment.getCommentId(), answerCommentId));
+
+        // 사용자에 대한 댓글을 찾을 수 없다면
+        if (!hasComment) throw new BusinessLogicException(ExceptionCode.COMMENT_NOT_FOUND);
+
+        AnswerComment answerComment = verifyExistsComment(answerCommentId);
+        answerComment.update(Optional.ofNullable(requestCommentDto.getContents()).orElse(answerComment.getContents()));
     }
 
 
     @Transactional
-    public void deleteComment(Long id) {
-        AnswerComment answerComment = verifyExistsComment(id);
+    public void deleteComment(Long answerCommentId, Object accountEmail) {
+        Account account = verifyExistsAccount(accountEmail);
+        AnswerComment answerComment = verifyExistsComment(answerCommentId);
+        boolean hasComment = account.getAnswerComments().stream().anyMatch(comment -> Objects.equals(comment.getCommentId(), answerCommentId));
+
+        if (!hasComment) throw new BusinessLogicException(ExceptionCode.COMMENT_NOT_FOUND);
+
 
         answerCommentRepository.delete(answerComment);
     }
 
-    private AnswerComment verifyExistsComment(Long id) {
-        Optional<AnswerComment> comment = answerCommentRepository.findById(id);
+    @Transactional(readOnly = true)
+    public Account verifyExistsAccount(Object principal) {
+        return accountRepository.findByEmail((String) principal)
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
+    }
 
-        return comment.orElseThrow(IllegalArgumentException::new);
+    @Transactional(readOnly = true)
+    public AnswerComment verifyExistsComment(Long id) {
+        return answerCommentRepository.findById(id)
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.COMMENT_NOT_FOUND));
     }
 
     @Transactional(readOnly = true)
@@ -76,9 +96,12 @@ public class AnswerCommentService {
 
             Page<AnswerCommentResponseDto> pageComments = answerCommentRepository.findByAnswer_AnswerId(answerId, PageRequest.of(page, size))
                 .map(comment -> AnswerCommentResponseDto.builder()
-                        .content(comment.getContents())
+                        .commentId(comment.getCommentId())
+                        .accountId(comment.getAccount().getAccountId())
                         .answerId(answer.getAnswerId())
+                        .contents(comment.getContents())
                         .displayName(comment.getAccount().getDisplayName())
+                        .email(comment.getAccount().getEmail())
                         .createAt(comment.getCreatedAt())
                         .modifiedAt(comment.getModifiedAt())
                         .build());
